refactor: post feature 세분화
