+++ src/features/post-detail/ui/PostDetailModal.tsx
