+++ src/features/post-view/model/usePosts.ts
@@ -1,5 +1,5 @@
 import { useState } from "react"
-import { createPostApi, deletePostApi, fetchPostsApi, updatePostApi } from "../../../entities/post/api"
+import { deletePostApi, fetchPostsApi } from "../../../entities/post/api"
 import { fetchUsersApi } from "../../../entities/user/api"
 import { Post } from "../../../entities/post/types"
 import { User } from "../../../entities/user/model/types"
@@ -10,25 +10,7 @@ export const usePosts = () => {
   const [showAddDialog, setShowAddDialog] = useState(false)
   const [selectedPost, setSelectedPost] = useState<Post | null>(null)
   const [showEditDialog, setShowEditDialog] = useState(false)
-  const [newPost, setNewPost] = useState({ title: "", body: "", userId: 1 })
   const [total, setTotal] = useState(0)
-
-  const addPost = async () => {
-    const data = await createPostApi(newPost)
-
-    setPosts([data, ...posts])
-    setShowAddDialog(false)
-    setNewPost({ title: "", body: "", userId: 1 })
-  }
-
-  const updatePost = async () => {
-    if (!selectedPost) return
-
-    const data = await updatePostApi(selectedPost)
-
-    setPosts(posts.map((post) => (post.id === data.id ? data : post)))
-    setShowEditDialog(false)
-  }
 
   const deletePost = (postId: number) => {
     deletePostApi(postId)
@@ -75,18 +57,15 @@ export const usePosts = () => {
 
   return {
     posts,
+    setPosts,
     showAddDialog,
     setShowAddDialog,
     selectedPost,
     setSelectedPost,
     showEditDialog,
     setShowEditDialog,
-    newPost,
-    setNewPost,
     total,
 
-    addPost,
-    updatePost,
     deletePost,
     getPostsByTag,
     getSearchedPosts,
