+++ src/features/post-view/ui/PostTable.tsx
