+++ src/pages/PostsManagerPage.tsx
@@ -2,19 +2,19 @@ import { useEffect, useState } from "react"
 import { Plus } from "lucide-react"
 import { Button, Card, CardContent, CardHeader, CardTitle, Loader, SearchInput } from "../shared/ui"
 import { useTags } from "../entities/tag/model"
-import { usePostParams } from "../features/post/model/usePostParams"
+import { usePostParams } from "../features/post-view/model/usePostParams"
 import { fetchUsersApi } from "../entities/user/api"
-import { PostDetailModal } from "../features/post/ui/PostDetailModal"
+import { PostDetailModal } from "../features/post-detail/ui/PostDetailModal"
 import { UserModal } from "../features/user/ui/UserModal"
-import { PostAddModal } from "../features/post/ui/PostAddModal"
-import { PostEditModal } from "../features/post/ui/PostEditModal"
+import { PostAddModal } from "../features/post-add/ui/PostAddModal"
 import { useUserContext } from "../shared/model/UserContext"
 import { usePostsContext } from "../shared/model/PostContext"
-import { PostTable } from "../features/post/ui/PostTable"
-import { Pagination } from "../features/post/ui/Pagination"
+import { PostTable } from "../features/post-view/ui/PostTable"
+import { Pagination } from "../features/post-view/ui/Pagination"
 import { SelectTag } from "../features/post-filter/ui/SelectTag"
 import { SelectSortStandard } from "../features/post-sort/ui/SelectSortStandard"
 import { SelectSortOrder } from "../features/post-sort/ui/SelectSortOrder"
+import { PostEditModal } from "../features/post-edit/ui/PostEditModal"
 
 const PostsManager = () => {
   const [loading, setLoading] = useState(false)
